server/requirements.txt
bidict==0.21.2
certifi==2020.12.5
chardet==3.0.4
click==7.1.2
dataclasses==0.6
dnspython==1.16.0
eventlet==0.30.1
filelock==3.0.12
Flask==1.1.2
Flask-Cors==3.0.9
Flask-SocketIO==5.0.1
future==0.18.2
greenlet==1.0.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.17.0
MarkupSafe==1.1.1
netifaces==0.10.6
nltk==3.5
numpy==1.19.4
packaging==20.7
pandas==1.2.2
#pkg-resources==0.0.0
protobuf==3.14.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-engineio==4.0.0
python-socketio==5.1.0
pytz==2021.1
regex==2020.11.13
requests==2.25.0
sacremoses==0.0.43
scikit-learn==0.23.2
scipy==1.5.4
sentence-transformers==0.3.9
sentencepiece==0.1.91
six==1.15.0
sklearn==0.0
threadpoolctl==2.1.0
tokenizers==0.9.3
torch==1.7.0
tqdm==4.54.1
transformers==3.5.1
typing-extensions==3.7.4.3
Unidecode==1.2.0
urllib3==1.26.2
Werkzeug==1.0.1
